Notes From 9/5 Workblock
Made a .txt file with notes on a few ideas Dale and Skyler were tossing around for Art things, including options for the 3rd music genre if we don't do Metal, options for the starting instrument for MC, MC's starting outfit, and Stage Layout.

Also made a folder to place these notes in.